chore: Use UUID instead of u128 for ids in tracing
ClickHouse doesn't have proper support for ULID, but it has for UUID.
So might as well benefit from it as it's more readable and Postgres also
stores UUID.
