+++ gateway/crates/federated-server/src/lib.rs
@@ -6,7 +6,6 @@
 
 pub use crate::config::{Config, TelemetryConfig};
 pub use error::Error;
-#[cfg(not(feature = "lambda"))]
 pub use server::GdnResponse;
 pub use server::{GraphFetchMethod, OtelReload, OtelTracing};
