+++ gateway/crates/federated-server/src/server.rs
@@ -10,11 +10,10 @@ mod state;
 mod trusted_documents_client;
 
 use grafbase_tracing::metrics::HasGraphqlErrors;
-#[cfg(not(feature = "lambda"))]
-pub use graph_updater::GdnResponse;
 
 pub use graph_fetch_method::GraphFetchMethod;
 pub use otel::{OtelReload, OtelTracing};
+use ulid::Ulid;
 
 use crate::config::Config;
 use crate::config::TlsConfig;
@@ -134,4 +133,22 @@ async fn bind(_: SocketAddr, path: &str, router: Router<()>, _: Option<TlsConfig
     lambda_http::run(app).await.expect("cannot start lambda http server");
 
     Ok(())
+}
+
+#[derive(Debug, serde::Serialize, serde::Deserialize)]
+#[allow(dead_code)]
+/// Response from the API containing graph information
+pub struct GdnResponse {
+    /// Account id of the owner of the referenced graph
+    pub account_id: Ulid,
+    /// The id of the graph
+    pub graph_id: Ulid,
+    /// The branch name
+    pub branch: String,
+    /// Grafbase id to uniquely identify the branch
+    pub branch_id: Ulid,
+    /// GraphQL SDL
+    pub sdl: String,
+    /// Current version's id generated by Grafbase
+    pub version_id: Ulid,
 }
