+++ gateway/crates/federated-server/src/server/graph_updater.rs
@@ -11,6 +11,8 @@ use tracing::Level;
 use ulid::Ulid;
 use url::Url;
 
+use super::GdnResponse;
+
 /// How often we poll updates from the schema registry.
 const TICK_INTERVAL: Duration = Duration::from_secs(10);
 
@@ -44,24 +46,6 @@ pub(super) struct GraphUpdater {
     current_id: Option<Ulid>,
     gateway_config: GatewayConfig,
     otel_reload: Option<oneshot::Sender<OtelReload>>,
-}
-
-#[derive(Debug, serde::Serialize, serde::Deserialize)]
-#[allow(dead_code)]
-/// Response from the API containing graph information
-pub struct GdnResponse {
-    /// Account id of the owner of the referenced graph
-    pub account_id: Ulid,
-    /// The id of the graph
-    pub graph_id: Ulid,
-    /// The branch name
-    pub branch: String,
-    /// Grafbase id to uniquely identify the branch
-    pub branch_id: Ulid,
-    /// GraphQL SDL
-    pub sdl: String,
-    /// Current version's id generated by Grafbase
-    pub version_id: Ulid,
 }
 
 impl GraphUpdater {
